# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CBS Core API
  version: 1.0.0
  summary: Kontrak HTTP inti CBS Core (BPR/BMT).
  description: |
    Spesifikasi kontrak API `/api/v1` untuk core banking CBS Core.

    ## Autentikasi
    Dua mekanisme dipakai bersamaan oleh klien resmi:

    - **Bearer JWT** (`Authorization: Bearer <access_token>`) — untuk integrasi
      non-browser dan pengujian.
    - **Cookie sesi httpOnly** (`cbs_access_token`) — dipakai web. Cookie
      `cbs_refresh_token` dipakai untuk memperbarui sesi.

    Endpoint publik (`/healthz`, `/api/v1/app-info`, `/api/v1/auth/login`,
    `/api/v1/auth/refresh`) tidak memerlukan autentikasi.

    Semua endpoint di belakang autentikasi juga melewati cakupan buku
    (`institution.book_scope`) dan cakupan unit organisasi (cabang/area/wilayah);
    data di luar cakupan aktor ditolak 403 atau disamarkan 404 pada pembacaan.

    ## CSRF
    Permintaan yang mengubah state (POST/PUT/DELETE) dari klien berbasis cookie
    wajib menyertakan header `X-CSRF-Token` yang nilainya sama dengan cookie
    `csrf_token` (pola double-submit). Klien Bearer murni dapat mengabaikannya.

    ## Amplop respons
    Semua respons sukses memakai amplop
    `{"success": true, "message": "...", "data": ..., "meta": ...}`.
    Respons galat memakai `{"success": false, "error": "..."}`.

    `data` daftar kosong selalu diserialkan sebagai `[]`/`{}`, tidak pernah `null`.
  contact:
    name: Tim CBS Core
servers:
  - url: http://localhost:8080
...
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /api/v1/ojk/placement-codes/{placementId}:
    parameters:
      - name: placementId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags: [Sistem]
      operationId: updateOJKPlacementCodes
      summary: Ubah sandi OJK penempatan
      description: >-
        Izin: `system:config`. Mengisi sandi referensi/inline OJK Form 05.00 per
        penempatan pada bank lain tanpa SQL: `ojk_kabupaten_code` (III, FK ke
        `ojk_kabupaten`), `ojk_hubungan_bank_code` (V, inline 12/20),
        `ojk_alasan_diblokir_code` (XI), `counterparty_cif` (XVI),
        `ojk_klasifikasi_aset_code` (XX), `blocked_amount` (X),
        `accrued_interest_receivable` (XIII), dan `accrued_interest_pending` (XIV).
        Semua bidang opsional bertipe string nullable: absen/null berarti tidak
        diubah, "" berarti dikosongkan (NULL), selain itu divalidasi lalu disimpan.
        Hanya bidang yang dikirim yang diubah; teraudit satu transaksi. Menolak 422
        bila sandi inline di luar himpunan, sandi referensi tidak ada di tabel, atau
        nominal negatif. Cakupan cabang aktor ditegakkan.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                ojk_kabupaten_code: { type: string, nullable: true }
                ojk_hubungan_bank_code: { type: string, nullable: true }
                ojk_alasan_diblokir_code: { type: string, nullable: true }
                counterparty_cif: { type: string, nullable: true }
                ojk_klasifikasi_aset_code: { type: string, nullable: true }
                blocked_amount: { type: string, nullable: true }
                accrued_interest_receivable: { type: string, nullable: true }
                accrued_interest_pending: { type: string, nullable: true }
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      tags: [Sistem]
      operationId: getOJKPlacementCodes
      summary: Baca sandi OJK penempatan tersimpan
      description: >-
        Izin: `system:config`. Mengembalikan nilai penempatan tersimpan. Sandi/teks/
        counterparty bertipe string atau null; nominal bertipe number atau null selama
        belum diisi (dibedakan dari nol). Cakupan cabang aktor ditegakkan: penempatan
        di luar cabang aktor dibalas 404 (disamarkan) mengikuti pembacaan kredit.
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token JWT dari `/api/v1/auth/login`.
    cookieAuth:
      type: apiKey
      in: cookie
      name: cbs_access_token
      description: |-
        Cookie sesi httpOnly. Nama dapat diubah lewat `CBS_ACCESS_COOKIE`.
        Permintaan yang mengubah state wajib menyertakan `X-CSRF-Token`.
...
  responses:
    OK:
      description: Berhasil.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
...
    BadRequest:
      description: Permintaan tidak dapat dibaca atau bidang wajib kosong.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Token hilang, kedaluwarsa, atau tidak sah.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: |-
        Izin tidak cukup, kata sandi kedaluwarsa, atau data di luar cakupan
        cabang/buku aktor.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Sumber daya tidak ditemukan (atau disamarkan pada pembacaan lintas cakupan).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    UnprocessableEntity:
      description: Permintaan terbaca tetapi gagal validasi/aturan bisnis.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: |-
        Kegagalan internal. Detail hanya tercatat di log; klien menerima pesan
        generik.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Envelope:
      type: object
      description: Amplop respons standar.
      properties:
        success:
          type: boolean
        message:
          type: string
        data: {}
        error:
          type: string
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          const: false
        error:
          type: string
      required: [success, error]
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
...